wip(econometrica): reproduce Figure 2
BUG: The maximizing MTRs are violating upper bound constraint!

Figure 1 and Figure 2 have a few differences:
- Figure 2 has a title
- Figure 2 uses information from `compute_bounds()`

To allow for these changes, I expanded the cases for when `mtroption` is
"max" or "min".
I modified the signature of `mtrs_and_weights()` to allow for `basis`
and `assumptions`.
`defaults` was renamed to `opts`.
I also removed the defaults, which forces me to be more explicit in
`toplevel.jl`.