wip(econometrica): reproduce Figure 4
The "IV Slope Indicator" is not a single IV-like estimand.
Each index specified in `:ivslopeind` corresponds to a single IV-like
estimand.

To accomodate this, I ran a for-loop across the different values in
`:ivslopeind`.
I also needed to return a vector when calling `legendtitle` and
`pathtitle`.
I could have returned a scalar that is specific to the IV-like estimand
corresponding to a particular index in `:ivslopeind`.
However, I liked the idea of having the input of `legendtitle` and
`pathtitle` be the IV-like object specified by the user.
I might change my mind on this.

Unlike before, the legend now uses `\mathbb{1}` to describe the IV Slope
Indicators.
Due to AT's preferred style of bb-fonts, I needed to run
`tlmgr install collection-fontsrecommended`.

TODO: should `legendtitle` and `pathtitle` return a scalar?
TODO: consider changing `ivslopeind` to `ivslopenp`
TODO: consider changing "indicator" to "nonparametric"