wip(econometrica): reproduce Figure 5
By default, `:saturated` can be thought to be true.
To use non-saturated set of IV-like specifications, we need to
explicitly assign the value `false` to `assumptions[:saturated]`.

Thus, when deciding whether to plot the `saturated` weights, it is not
enough to check whether `assumptions` has the `:saturated` key. We must
also check if the value of this key is true.
I added a similar check to the other IV-like estimands.

Unlike the other IVLike structs, the `make_slist` IVLike struct iterates
over `eachrow(dgp.suppZ)`. As a result, the s function is a function of
d \in {0, 1} and z \in {[dgp.suppZ[1]], [dgp.suppZ[2]], ...}.
That is, the second argument is a vector of instrument values.
The other s functions have scalars as the second argument because they
were coded with the understanding that there is a single instrument.

This distinction forces us to compute the saturated weights differently
from the other weights.

TODO: ensure all the s functions have the same signature.