fix(mtr): break into segments to avoid vertical lines
Before this commit, we would be drawing a single path for each MTR.
As a result, there would be a vertical line at the points of
discontinuity.
To get rid of this, I modified `df_to_coordinates` so that it doesn't
return a single string of coordinates.
Instead, it returns a vector of strings, each of which are the
coordinates for a particular path.
Each string corresponds to its own path, i.e., we are now drawing
multiple paths for each MTR instead of a single path.
The tikz-template.tex has been changed accordingly.

This begs the question:
If we know the MTR is piecewise constant (i.e., the basis is a
constant spline basis), could we rely on the same code as the weights
for the IV-like estimands and target parameters?

TODO: combine `df_to_coordinates` for MTRs and for weights?
TODO: remove duplication in `mtrs_and_weights`
NEXT: code everything!