wip(econometrica): reproduce Figure 8
Figure 8 relies on a different template than the other figures.

This figure required me to increase the resolution, i.e., instead of
rounding the coordinates to 3 decimal places, I needed to round to 4.
Otherwise, the curves look jagged.
Furthermore, I ended up allowing the discontinuity tolerance to be a
parameter in `df_to_coordinates()`.