wip(review): reproduce Figure 3
I create a tikz-template file that is hopefully general enough to
reproduce Figures 3, 10, and 11.
The major challenge was to take care of the comma-separated values in
the `extra x ticks` and `extra x tick labels`.
I tried iterating using Mustache, but it would leave behind a trailing
comma that caused (harmless) errors when compiling.
I ultimately decided to feed a string of comma-separated values.

I also had to deal with NaN values when evaluating the bounds.

To get the data for the curves, I relied on `compute_bounds` with the
dgp bases and a saturated set of IV-like estimands because LATEs are
point-identified.
However, it might be better to "manually" compute the LATEs w/ DGP info.

TODO: consider manually computing LATEs
TODO: get rid of duplication
  - within for-loop
  - across different curves