wip(review): reproduce Figure 4
This is the first figure that uses a plotting function that is common
between both projects.
To access the correct template, I made a variable called `project` in
the global scope.

There are a few differences between Figure 4 in the Review paper and the
other figures in the Econometrica paper.
First, the Review paper makes use of linetypes instead of marks.
  -> I introduced `linetypes` in `opts`
Second, the Review paper has more to the title after the bounds.
  -> I introduced `titlesuffix` in `opts`
The template files for the Review and Econometrica paper were adapted to
make use of the relevant variables when necessary.

Figure 4 also uses the ATT as the target parameter and the TSLS Slope
Indicator as an IV-like estimand.
They were straightforward to accomodate.
The only wrinkle is that the ATT requires information from dgp::DGP to
compute the average weights.
So, I needed to add dgp::DGP in `compute_average_weights`.
Everything should still be backwards-compatible.