wip(review): reproduce Figure 6
I based the template off of `src/python/tikz-template-kbounds.tex`.
The template is rather flexible -- it requires me to specify `opts` for
each `\addplot`.
This allowed me to copy the options and the code verbatim.
This was especially useful when shifting the cycle to allow different
`\addplot` to share the same color.