feat(plotting): return bounds and fix coefficients
To match Figure 4 of MST (2018), I need to fix the coefficient
associated with the first "piece" of the maximizing MTR for d = 1.
Hence, after introducing `fixdf` in `compute_bounds` in
`MarginalTreatmentEffectsWithMultipleInstruments.jl`, I expanded
`mtrs_and_weights` to use this feature.

To ensure that the bounds don't change as a result of this additional
constraint, I want to return the bounds so I can check the results
later.