refactor: avoid duplication; store tex files and compile all at once
The end of each function in `mst2018econometrica.jl` and
`mt2018review.jl` compiles the tex file if the `compile` argument is
true.

To avoid duplication, I'll instead return the name of the tex file in
each of these functions.
I'll store these file names in `texfiles`, a global variable defined in
`menu()`.
Then, at the end of `menu()`, I will compile the tex files.

This has the added benefit of only using the `compile` argument in
`menu()`.

For `run_np_ivnps()`, two tex files are produced so a tuple is returned.
To add these files to `texfiles`, I rely on the splat operator `...`.