feat: remove MarginalTreatmentEffectsWithMultipleInstruments.jl
I copied the code from
MarginalTreatmentEffectsWithMultipleInstruments.jl to this package.
This removes one dependency for the package and houses the replication
code for all three papers under one unified package.

The replication for the MTW (2021) paper doesn't use Mustache.
Instead, the strategy is to store the data in CSVs and have pgfplots
read in the data to create the plots.
The benefit of this approach is that we can use these CSVs for testing.
Switching to Mustache makes it difficult to use these CSVs for testing.

The tests fail on my system.
I don't remember them failing before, but the failures are quite minor.
Instead of using `isequal`, using `isapprox` returns successes.
Note that `isapprox` doesn't work for strings, so for the PRTE
Misspecification test, I only check the Float64 columns with `isapprox`,
not the String columns.

There's a lot of exporting from the original package.
I think I should avoid exporting them with this new package and only
use `menu()` for exporting.
I need to check with a-torgivitsky.

To be consistent with the existing structure of
MarginalTreatmentEffects.jl, I should move the code that is specific to
MTW (2021) to mtw2021econometrics.jl.
For example, the DGPs for the MTW project are stored in dgp.jl.
I should instead move them to mtw2021econometrics.jl.

Q(a-torgovitsky): should I use Mustache for the MTW (2021) project?
Q(a-torgovitsky): what should I export?
TODO(omkarakatta): move project-specific code to mtw2021econometrics.jl