feat(toplevel): replicate all projects
I abstracted the tag generation so I won't have to do it for each
project. I also made a distinction between an empty tag and a tag that is
`nothing`. This allowed me to distinguish between (1) using date as the
tag and (2) asking user for input.

I used `@assert` instead of `@error` to check user input.