refactor(econometrics): save file names and compile at the end
Before this commit, we would be compiling the tex files within each
function of `mtw2021econometrics.jl`.
After this commit, we would only collect the tex files in `texfiles` and
compile at the end.

This strategy avoids duplication, and it further emphasizes the use of
`menu` as the main function for users to interface with.